Move test-data into tests
